requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

python-dotenv==1.0.1

supabase==2.9.1
httpx[http2]==0.27.2
websockets==12.0

gotrue==2.9.3
postgrest==0.17.1
realtime==2.0.6
storage3==0.7.7

pydantic==2.9.2
pydantic-settings==2.6.1
typing_extensions==4.12.2

zeep
requests